Just rearranging file structure
moving things around and cleaning up a bit
