--- A2_AppProject/A2_AppProject/A2_AppProject/AddHabit.xaml
@@ -1,19 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="A2_AppProject.AddHabit"
-             Title="Add Habit">
-    
-    <ContentPage.Content>
-
-        <StackLayout>
-
-            <Label Text="Placeholder for Add Habit"
-                VerticalOptions="CenterAndExpand" HorizontalOptions="CenterAndExpand" />
-
-            <Image Source="egAddHabit.png" />
-
-        </StackLayout>
-
-    </ContentPage.Content>
-</ContentPage>
