+++ A2_AppProject/A2_AppProject/A2_AppProject/Pages/HabitsView.xaml
@@ -1,0 +1,17 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="A2_AppProject.HabitsView">
+    <ContentPage.Content>
+
+        <StackLayout>
+
+            <Label Text="Placeholder for Add Habit"
+                VerticalOptions="CenterAndExpand" HorizontalOptions="CenterAndExpand" />
+
+            <Image Source="egAddHabit.png" />
+
+        </StackLayout>
+
+    </ContentPage.Content>
+</ContentPage>
