+++ A2_AppProject/A2_AppProject/A2_AppProject/Pages/HabitsView.xaml.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using Xamarin.Forms;
+using Xamarin.Forms.Xaml;
+
+namespace A2_AppProject
+{
+	[XamlCompilation(XamlCompilationOptions.Compile)]
+	public partial class HabitsView : ContentPage
+	{
+		public HabitsView ()
+		{
+			InitializeComponent ();
+		}
+	}
+}
