+++ A2_AppProject/A2_AppProject/A2_AppProject/Pages/MonthView.xaml
@@ -1,0 +1,18 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="A2_AppProject.MonthView"
+             Title="Month View">
+    
+    <ContentPage.Content>
+        <StackLayout>
+            
+            <Label Text="Placeholder for Month View"
+                VerticalOptions="CenterAndExpand" HorizontalOptions="CenterAndExpand" />
+
+            <Image Source="egMonthView.png" />
+
+        </StackLayout>
+    </ContentPage.Content>
+    
+</ContentPage>
