+++ A2_AppProject/A2_AppProject/A2_AppProject/Pages/TasksView.xaml
@@ -1,0 +1,19 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="A2_AppProject.Tasks"
+             Title="Tasks">
+    
+    <ContentPage.Content>
+
+        <StackLayout>
+
+            <Label Text="Placeholder for Tasks"
+                VerticalOptions="CenterAndExpand" HorizontalOptions="CenterAndExpand" />
+
+            <Image Source="egAddToDo.png" />
+
+        </StackLayout>
+
+    </ContentPage.Content>
+</ContentPage>
