--- A2_AppProject/A2_AppProject/A2_AppProject/TodayView.xaml.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using Xamarin.Forms;
-using Xamarin.Forms.Xaml;
-
-namespace A2_AppProject
-{
-	[XamlCompilation(XamlCompilationOptions.Compile)]
-	public partial class TodayView : ContentPage
-	{
-		public TodayView ()
-		{
-			InitializeComponent ();
-		}
-	}
-}
